# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_ProTrack-Solutions_protrack-api_internal_meta_whatsapp_domain.CompanyWhatsAppConfig:
    properties:
      companyID:
        type: string
      createdAt:
        type: string
      displayPhoneNumber:
        type: string
      id:
        type: string
      isActive:
        type: boolean
      mode:
        $ref: '#/definitions/github_com_ProTrack-Solutions_protrack-api_internal_meta_whatsapp_domain.WhatsAppMode'
      monthlyMessageAllowance:
        type: integer
      phoneNumberID:
        type: string
      updatedAt:
        type: string
      wabaid:
        type: string
    type: object
  github_com_ProTrack-Solutions_protrack-api_internal_meta_whatsapp_domain.Message:
    properties:
      category:
        $ref: '#/definitions/github_com_ProTrack-Solutions_protrack-api_internal_meta_whatsapp_domain.MessageCategory'
      companyID:
        type: string
      createdAt:
        type: string
      deliveredAt:
        type: string
      estimatedCostCents:
        format: int32
        type: integer
      failureCode:
        format: int32
        type: integer
      failureReason:
        type: string
      id:
        type: string
      metaMessageID:
        type: string
      recipientPhone:
        type: string
      sentAt:
        type: string
      status:
        $ref: '#/definitions/github_com_ProTrack-Solutions_protrack-api_internal_meta_whatsapp_domain.MessageStatus'
      templateID:
        type: string
    type: object
  github_com_ProTrack-Solutions_protrack-api_internal_meta_whatsapp_domain.MessageCategory:
    enum:
    - marketing
    - utility
    - authentication
    - service
    type: string
    x-enum-varnames:
    - CategoryMarketing
    - CategoryUtility
    - CategoryAuthentication
    - CategoryService
  github_com_ProTrack-Solutions_protrack-api_internal_meta_whatsapp_domain.MessageStatus:
    enum:
    - queued
    - sent
    - delivered
    - read
    - failed
    type: string
    x-enum-varnames:
    - StatusQueued
    - StatusSent
    - StatusDelivered
    - StatusRead
    - StatusFailed
  github_com_ProTrack-Solutions_protrack-api_internal_meta_whatsapp_domain.SendMessageRequest:
    properties:
      category:
        $ref: '#/definitions/github_com_ProTrack-Solutions_protrack-api_internal_meta_whatsapp_domain.MessageCategory'
      header_variables:
        additionalProperties:
          type: string
        description: chaves "1", "2"... posição no cabeçalho
        type: object
      language_code:
        type: string
      recipient_phone:
        type: string
      template_name:
        type: string
      variables:
        additionalProperties:
          type: string
        description: chaves "1", "2"... posição no corpo
        type: object
    type: object
  github_com_ProTrack-Solutions_protrack-api_internal_meta_whatsapp_domain.Template:
    properties:
      body_text:
        type: string
      category:
        $ref: '#/definitions/github_com_ProTrack-Solutions_protrack-api_internal_meta_whatsapp_domain.MessageCategory'
      created_at:
        type: string
      id:
        type: string
      is_platform_shared_eligible:
        type: boolean
      language_code:
        type: string
      meta_approval_status:
        description: pending | approved | rejected | paused
        type: string
      meta_template_name:
        type: string
      updated_at:
        type: string
      variables:
        items:
          type: string
        type: array
    type: object
  github_com_ProTrack-Solutions_protrack-api_internal_meta_whatsapp_domain.UpsertCompanyConfigRequest:
    properties:
      access_token:
        type: string
      display_phone_number:
        type: string
      is_active:
        type: boolean
      mode:
        $ref: '#/definitions/github_com_ProTrack-Solutions_protrack-api_internal_meta_whatsapp_domain.WhatsAppMode'
      monthly_message_allowance:
        type: integer
      phone_number_id:
        type: string
      waba_id:
        type: string
    type: object
  github_com_ProTrack-Solutions_protrack-api_internal_meta_whatsapp_domain.WhatsAppMode:
    enum:
    - platform_shared
    - own_waba
    type: string
    x-enum-varnames:
    - WhatsAppModePlatformShared
    - WhatsAppModeOwnWABA
...
  github_com_ProTrack-Solutions_protrack-api_internal_payment_methods_domain.CreatePaymentMethodRequest:
    properties:
      company_id:
        type: string
      name:
        type: string
      type:
        type: string
    type: object
  github_com_ProTrack-Solutions_protrack-api_internal_payment_methods_domain.GetPaymentMethodsStatsResponse:
    properties:
      payment_method:
        type: string
      percentage_method:
        type: number
    type: object
  github_com_ProTrack-Solutions_protrack-api_internal_payment_methods_domain.PaymentMethodResponse:
    properties:
      company_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      type: {}
      updated_at:
        type: string
    type: object
  github_com_ProTrack-Solutions_protrack-api_internal_payment_methods_domain.TogglePaymentMethodActiveRequest:
    properties:
      id:
        type: string
      is_active:
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact:
    email: suporte@protrack.com
    name: Suporte ProTrack
  description: API de gerenciamento financeiro e comercial ProTrack Solutions.
  termsOfService: http://swagger.io/terms/
  title: ProTrack API
  version: "1.0"
paths:
...
  /meta-whatsapp/config:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ProTrack-Solutions_protrack-api_internal_meta_whatsapp_domain.CompanyWhatsAppConfig'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Busca a configuração de WhatsApp da empresa
      tags:
      - meta-whatsapp
  /meta-whatsapp/list-templates:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_ProTrack-Solutions_protrack-api_internal_meta_whatsapp_domain.Template'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Lista os templates de WhatsApp aprovados
      tags:
      - meta-whatsapp
  /meta-whatsapp/send-message:
    post:
      consumes:
      - application/json
      description: Dispara uma mensagem de template aprovado para o número informado,
        respeitando o modo de envio (platform_shared ou own_waba) configurado para
        a empresa
      parameters:
      - description: Dados da mensagem
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_ProTrack-Solutions_protrack-api_internal_meta_whatsapp_domain.SendMessageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ProTrack-Solutions_protrack-api_internal_meta_whatsapp_domain.Message'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Envia uma mensagem de template via WhatsApp
      tags:
      - meta-whatsapp
  /meta-whatsapp/upsert-config:
    post:
      consumes:
      - application/json
      parameters:
      - description: Configuração de WhatsApp
        in: body
        name: config
        required: true
        schema:
          $ref: '#/definitions/github_com_ProTrack-Solutions_protrack-api_internal_meta_whatsapp_domain.UpsertCompanyConfigRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ProTrack-Solutions_protrack-api_internal_meta_whatsapp_domain.CompanyWhatsAppConfig'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Cria ou atualiza a configuração de WhatsApp da empresa
      tags:
      - meta-whatsapp
  /meta-whatsapp/webhook:
    get:
      description: 'Endpoint de handshake exigido pela Meta ao configurar o webhook:
        valida hub.verify_token e ecoa hub.challenge quando hub.mode é "subscribe"'
      parameters:
      - description: Modo do handshake (subscribe)
        in: query
        name: hub.mode
        required: true
        type: string
      - description: Token de verificação configurado na Meta
        in: query
        name: hub.verify_token
        required: true
        type: string
      - description: Valor a ser ecoado quando a verificação for bem-sucedida
        in: query
        name: hub.challenge
        required: true
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: challenge
          schema:
            type: string
        "403":
          description: Forbidden
      summary: Verifica o webhook do WhatsApp junto à Meta
      tags:
      - meta-whatsapp
    post:
      consumes:
      - application/json
      description: Endpoint chamado pela Meta para notificar status de mensagens e
        eventos da conta. A assinatura da requisição é validada pelo MetaWhatsappMiddleware
        via header X-Hub-Signature-256. Sempre responde 200 para evitar reenvios da
        Meta, mesmo em caso de falha no processamento (o erro é registrado no Discord)
      responses:
        "200":
          description: OK
      summary: Recebe eventos do webhook do WhatsApp
      tags:
      - meta-whatsapp
...
  /payment-methods:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_ProTrack-Solutions_protrack-api_internal_payment_methods_domain.PaymentMethodResponse'
            type: array
      security:
      - BearerAuth: []
      summary: Lista métodos de pagamento da empresa
      tags:
      - payment-methods
    post:
      consumes:
      - application/json
      parameters:
      - description: Método de pagamento
        in: body
        name: method
        required: true
        schema:
          $ref: '#/definitions/github_com_ProTrack-Solutions_protrack-api_internal_payment_methods_domain.CreatePaymentMethodRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
      security:
      - BearerAuth: []
      summary: Cria um método de pagamento
      tags:
      - payment-methods
  /payment-methods/{id}:
    get:
      parameters:
      - description: ID do método
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ProTrack-Solutions_protrack-api_internal_payment_methods_domain.PaymentMethodResponse'
      security:
      - BearerAuth: []
      summary: Busca método de pagamento por ID
      tags:
      - payment-methods
    put:
      consumes:
      - application/json
      parameters:
      - description: ID do método
        in: path
        name: id
        required: true
        type: string
      - description: Status
        in: body
        name: toggle
        required: true
        schema:
          $ref: '#/definitions/github_com_ProTrack-Solutions_protrack-api_internal_payment_methods_domain.TogglePaymentMethodActiveRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - BearerAuth: []
      summary: Ativa ou desativa um método de pagamento
      tags:
      - payment-methods
  /payment-methods/is-active:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_ProTrack-Solutions_protrack-api_internal_payment_methods_domain.PaymentMethodResponse'
            type: array
      security:
      - BearerAuth: []
      summary: Lista métodos de pagamento ativos
      tags:
      - payment-methods
  /payment-methods/stats:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ProTrack-Solutions_protrack-api_internal_payment_methods_domain.GetPaymentMethodsStatsResponse'
      security:
      - BearerAuth: []
      summary: Estatísticas dos métodos de pagamento
      tags:
      - payment-methods
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"